ExtractOp: Preserve container during iteration
It looks like temporaries in the range-expression get destroyed after
it's evaluated. In this case the range-expression is a call to
`Optional::operator*` that returns a reference to the Optional's inner
value, but the Optional itself is a temporary.

Fix this by assigning the value inside the optional to a variable.